Update Accordion.html
BUGFIX > When adding a accordion in TYPO3 backend, for each panel added, a field "title" AND a field "class Css" is purposed. However The Css field is never taken in account in the template accordion.html. 

This bugfix purpose to ad the {panel.panel.class} in the accordion template to make the css field in backend usable. (the same modification have to be done for type of fluidcontent TAB).

Best regards. 
Vincent (first bugfix of my life)